First implementation of category-matching stash code.
Eradicated all usage of IHCategory; "Categories" are now a dictionary mapping ItemCat enum values to an item-matching function (which is really all the IHCategory class ever did, anyway). This made it easier match categories in a chest to those in the inventory as we were key-ing on an enum rather than an object.

The code nearly works, too.  The only mystery is why an empty chest still matches the "Other" category.