Changed to use ItemID consts....also more categories.
Actually, a crazy amount of changes, mostly centering around my stupid mistake of not realizing that all the trouble of figuring out the tileIDs WAS ALREADY DONE FOR ME...don't know why I didn't think that the game itself would need to know those somehow.... Anyway, also moved the lock tex initialization to the Mod Base; moved the item extensions to CategoryDef and added some more; explored using the IsItemAllowed override of the ItemSlot class, but decided to leave that for later. Of course, more re-jiggering of category-matching and sorting rules.  Created a separate array of ItemCat IDs in order to check for item categories in a different order than the final sort order. And i think some other stuff i don't want to type it all here thats what git diff is for right well im going to sleep now.