some fixes. now pretty much works (even guide), w/ a few issues:
Shift-moving back to inventory jumbles stacks around. I think it should look for an empty slot *first*, then stack if it can't find any. This will lead to the more natural feeling of moving a "stack" as a single item.

Shift-Clicking on the coin or ammo slots moves one of the first 4 items from the hotbar...which was honestly a bit amusing at first.

Shift-clicking on accessory or item slots now does nothing if the inventory is open. I had intended for it to *not* override the default left-click action on those slots, but must have messed it up somehow. (Shift-right click on this slots is also non-functional, for that matter).

Finally, this breaks shift-to-buy from a merchant (actually, now that I think about it, did that ever work in the first place?), though shift-to-sell still works. This may be a complicated one to fix, in which case it probably isn't worth it. Will have to look into it.

But there are some nice improvements and fixes, as well: limiting the guide slot to only accept materials fixed those "undefined" crash errors, and that's now a handy little feature; plus, *all* shift-move actions now update the recipe list, which didn't happen in the other mod.