added keybind reminder to tooltip; some docstrings Issue: The key that displayed on the tooltips for Restock & SmartStash only shows the 2nd key (e.g. 'X', 'Q', '=', etc.) and NOT the modifier (i.e. Shift) that is required to activate those functions; thus when the player presses the indicated key, the wrong action will occur. To further complicate things, when the player DOES press shift, these buttons switch to the original game actions (quickstack & deposit-all), meaning that pressing those keys now will AGAIN perform a different action than what is displayed. This is all very confusing.
It seems I had not thought through the coupling of the shortcut-key component of the mod to the graphical component.... Separately, each performs exactly how I want it; but try to tie them together and it all falls apart. I'll either need to figure out a way to reconcile the two, or scrap the whole idea.