created ButtonBase subclass for the text btns
All the tweaks and checks and bizarre adjustments I was making to the base class (and various other places in the project) were getting ridiculous, so I pulled all the weird and specific stuff into a subclass, and left out some things that these buttons don't need, to reduce processing time.  They now seem to look and behave pretty much exactly like the vanilla buttons, but with all our extra bells and whistles.

One issue remaining is that visual indication of "locked" status does not stay in sync in between the text and the graphical buttons; the actual locking mechanism still works fine, though.