Hid non-mod text; all functionality seems good
The "Cancel" text is still too high when in Chest-Renaming mode, overlapping with the Save button. But all the buttons seem to work just like normal, so this is just an easily-fixed cosmetic issue. Another cosmetic issue may be that we should add some spacing between the buttons in icon-replacer mode, rather than having them bumping up against each other. Also see if the sort, etc. buttons can be added in.