Refactoring in eternal quest for efficiency
Previous methods of creating/adding-functionality-to buttons all had serious flaws with regards to their generic applicability and the matter in which information was transferred around the mod. This new approach is meant to establish a clear stacking-hierarchy of containers:
  - On the top of the hierarchy is the ButtonLayer, the only one of its kind in a button stack (though there may be other buttonlayers defined for displaying under different game conditions) and defines the area on the screen where the buttons will appear.
  - The ButtonLayer contains at least one but usually multiple ButtonBases, each of which outlines a space for a single button within its parent layer.
  - On the bottom of the stack is the new CoreButton class, intended to replace both IHButton & ButtonState.

The CoreButton is a very general class containing little actual implementation of behavior, with most of its attributes being defined at runtime. The new approach here, and the attempt to curb the scaling problem, is the introduction of a hook & plugin/service-based infrastructure.

Each CoreButton has a ButtonHooks object that contains fields for all the available drawing & behavior hooks, though none of these are filled in at construction. Each Corebutton, no matter its intended function, will be mostly identical to every other CoreButton, and will do pretty much nothing by itself.

Functionality is added to the button by subscribing to "services", each of which is a separate class, a subclass of the abstract ButtonService, and should be defined to add a very specific function to button or change a single aspect of how it is displayed. Services contain a ButtonHooks object of their own, and can assign custom Delegates to the fields of this object.  For each hook they assign, they will register that hook with their subscribed client; what this means is that, for example, when the Draw command for the button calls the "preDraw" hook for a CoreButton, that button will in turn call the "preDraw" hook of each subscribed ButtonService that has registered said hook with the Button. Services can even add and remove hooks dynamically as the game runs and the player takes actions.

In this way, every button that needs a certain functionality will simply subscribe to the Service for that function and receive all the associated hooks. Hopefully it'll get rid of the huge headache that defining and switching states and contexts was before this.

TODO: Still need to write and test more Services. ButtonBase needs to be rewritten (in the placeholder "ButtonRebase" file). I was also in the middle of writing a kind of "fluid" interface or whatever in ButtonFactory to make creating a button like describing it, but I think that that will need to either be rewritten or just scrapped.  All the button-creation code will need to be redone. Why do I do this to myself?