enabled tooltip-drawing. Hopefully.
The tooltip to be drawn is now queued in a global stack and drawn
during the ModInterface.PostDrawInventory() hook. I'm hoping that
drawing it after the inventory will prevent the previous problem of
tooltips being drawn underneath other UI elements.

I don't think there are any non-inventory elements the tips may overlap
with, but if so, the call may need to be moved to PostDrawInterface().