simplified a number of common calls
Since we're storing a reference to the current IHPlayer instance,
there's no need to have TAPI figure it out from the ModPlayer list. I
think this remedied some of the lag that occurs when first using one of
the sorting or movement functions, but I'm not sure. Still a benefit
overall.