Sound feedback for C. Stacks; enhance readability
Specifically, rename a few variables to better indicate what they're
meant to represent, and improve the documentation on a few functions.

Also, StackItems(), which is called by ConsolidateStacks(), now plays
the ItemMoved Sound if it changes the stack value of any item. This is
mainly to add some sort of feedback to Clean Stacks when it is
activated manually (was previously silent). With new SoundManager
class, this sound should only play once per frame even though it may be
called multiple times here, many times in Sort(), and perhaps
elsewhere.