Add SoundManager class to...manage sounds.
More specifically, any time a sound is played in the mod with
.Play() (e.g. Sound.ItemMoved.Play(); ), it is actually being placed
into the SoundManager's queue of sounds. This "queue" is actually a
set (HashSet<Sound>), meaning that a new sound will only be added if it
is not already in the queue. At the end of the frame (in
ModInterface.PostDrawInventory), all the sounds in the SoundManager
queue are played and the queue is emptied.

This system prevents certain sounds from being played perhaps dozens
of time per frame. The most egregious is the ItemMoved sound, set
to play every time an item stack's value changes, every time an item
moves slots, etc., all of which can and often do happen multiple times
with any given operation of this mod. Now, no matter how many item-moved
events happen, the sound will only play once that frame.

In addition to taking a noticeable amount of stress off the CPU, this also
(in my opinion) reduces the aural dissonance that comes from the sound
playing numerous times in succession. And, due to how many times per second
a frame occurs in Terraria, there is no noticeable delay between the
action and the sound effect.